Trains a GELNET model

Trains a model on the definition constructed by gelnet()

gelnet_train(modeldef, max_iter = 100L, eps = 1e-05, silent = FALSE,
  verbose = FALSE)

Arguments

modeldef
model definition constructed through gelnet() arithmetic

max_iter
maximum number of iterations

eps
convergence precision

silent
set to TRUE to suppress run-time output to stdout; overrides verbose (default: FALSE)

verbose
set to TRUE to see extra output; is overridden by silent (default: FALSE)

Value

A GELNET model, expressed as a list with two elements:

w

p-by-1 vector of p model weights

b

scalar, bias term for the linear model (omitted for one-class models)

Details

The training is performed through cyclical coordinate descent, and the optimization is terminated after the desired tolerance is achieved or after a maximum number of iterations.